+++ swiyu-core-business-service-registry-identifier/src/main/java/ch/admin/bj/swiyu/registry/identifier/domain/IdentifierDatastoreEntity.java
@@ -10,12 +10,14 @@ import jakarta.persistence.*;
 import jakarta.validation.Valid;
 import java.util.UUID;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
 @Getter
 @EntityListeners(AuditingEntityListener.class)
 @Table(name = "datastore_entity")
+@NoArgsConstructor // JPA required
 public class IdentifierDatastoreEntity {
 
     @Embedded
@@ -23,7 +25,6 @@ public class IdentifierDatastoreEntity {
     private final AuditMetadata auditMetadata = new AuditMetadata();
 
     @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
     private UUID id;
 
     @Version
@@ -34,7 +35,8 @@ public class IdentifierDatastoreEntity {
     @Enumerated(EnumType.STRING)
     private DatastoreStatus status;
 
-    public IdentifierDatastoreEntity() {
+    public IdentifierDatastoreEntity(UUID id) {
+        this.id = id;
         this.status = DatastoreStatus.SETUP;
     }