+++ swiyu-core-business-service-registry-identifier/src/test/java/ch/admin/bj/swiyu/registry/identifier/test/IdentifierRegistryTestData.java
@@ -4,6 +4,7 @@ import ch.admin.bj.swiyu.registry.identifier.domain.DatastoreStatus;
 import ch.admin.bj.swiyu.registry.identifier.domain.DidEntity;
 import ch.admin.bj.swiyu.registry.identifier.domain.DidType;
 import ch.admin.bj.swiyu.registry.identifier.domain.IdentifierDatastoreEntity;
+import java.util.UUID;
 import lombok.experimental.UtilityClass;
 
 @UtilityClass
@@ -18,7 +19,7 @@ public class IdentifierRegistryTestData {
     }
 
     public static IdentifierDatastoreEntity datastoreEntity(DatastoreStatus status) {
-        var entity = new IdentifierDatastoreEntity();
+        var entity = new IdentifierDatastoreEntity(UUID.randomUUID());
         entity.changeStatus(status);
         return entity;
     }