+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 3.40.0
+
+### Changed
+
+- Identifier ID generation now explicit
+
+### Added
+
+- Support for local E2E tests
+
+## 3.39.7
+
+### Fixed
+
+- Fix identifier update returning 500 instead of 400 for a did log with an unrecognized DID method
+
 ## 3.39.6
 
 ### Changed
@@ -36,7 +52,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add declaration of intent templates in it,fr,en,de for actor type gov and biz
 
-
 ## 3.39.1
 
 ### Changed
@@ -47,7 +62,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add multi-language entity names for business partner V2 responses and persist new trust onboarding organization name translations.
+- Add multi-language entity names for business partner V2 responses and persist new trust onboarding organization name
+  translations.
 
 ## 3.38.2